// src/lib/interfaces/link.ts
export interface Links {
	folderId: string;
	folderName: string;
	folderColor: string;
	linksId: string;
	title: string;
	shortCode: string;
	destinationUrl: string;
	expireAt: Date;
	isActive: boolean;
	noOfClicks: number;
	createdAt: Date;
	updatedAt: Date;
	protected: boolean;
	imageData: string;
	format: string;
	size: number;
}